fix #229
Pad the item inserted into the minibuffer with a space on both sides.
This fix helps in the following cases:

1. User selects one or more items and starts the minibuffer.
2. Minibuffer is already started and user clicks an item.
3. Minibuffer is already started and ends with a space and user drops
   some items from the same or another window onto the minibuffer.

Comment:
In (3) rox can't directly change dropped items because the drop action
is built into the entry widget therefore rox can't control it. However,
by adding a space _after_ the items that rox can control (1 and 2),
also the usability of drop actions is improved.